#185 Siirretty v1 mockit omaan tiedostoonsa
